+++ lib/ui/widgets/buttons/secondary_light_button/index.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 
-class SecondaryLightRoundedButton extends StatelessWidget {
+class SecondaryLightButton extends StatelessWidget {
   final void Function() onPressed;
   final Widget child;
 
-  const SecondaryLightRoundedButton({
+  const SecondaryLightButton({
     Key key,
     @required this.onPressed,
     @required this.child,
@@ -13,8 +13,10 @@ class SecondaryLightRoundedButton extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     Color color = Colors.white;
+    Color background = Theme.of(context).colorScheme.primary;
     if (Theme.of(context).brightness == Brightness.dark) {
       color = Theme.of(context).colorScheme.primary;
+      background = Colors.transparent;
     }
 
     return FlatButton(
@@ -25,7 +27,7 @@ class SecondaryLightRoundedButton extends StatelessWidget {
         borderRadius: BorderRadius.circular(4),
         side: BorderSide(color: color, width: 0.5),
       ),
-      color: Colors.transparent,
+      color: background,
       textColor: color,
     );
   }