+++ lib/usecases/posts/usecase_delete_post.dart
@@ -1,0 +1,25 @@
+import 'package:meta/meta.dart';
+import 'package:mooncake/usecases/usecases.dart';
+import 'package:mooncake/entities/entities.dart';
+
+/// Allows to delete a locally stored post.
+class DeletePostUseCase {
+  final PostsRepository _postsRepository;
+
+  DeletePostUseCase({
+    @required PostsRepository postsRepository,
+  })  : assert(postsRepository != null),
+        _postsRepository = postsRepository;
+
+  /// Deletes the locally stored post if it hasn't been posted.
+  Future<void> delete(Post post) async {
+    List<PostStatusValue> canDeleteStatus = [
+      PostStatusValue.STORED_LOCALLY,
+      PostStatusValue.ERRORED
+    ];
+
+    if (canDeleteStatus.contains(post.status.value)) {
+      return _postsRepository.deletePost(post);
+    }
+  }
+}